Add expressions to config
Parse and execute expressions on module fields. This is a first step
torwards dynamic modules :
* Per host rate limiting / sampling
* Arbitrary filtering
* Sink split: use one kafka queue per arbitrary grouping value for
example
* ...

The expression languages uses Google's CEL and is statically typed to
validate the configuration before handing requests.

The modules are not designed to react to value change for their
parameters that depend on expressions. This is by design as some modules
such as rate limiting or sampling need stable parameter, a
control.group_by module will be added in an other PR to dynamically
spawn modules based on arbitrary values.